+++ mp-opt-model/lib/@opt_model/parse_soln.m
@@ -46,6 +46,10 @@ if nargin < 2
     stash = false;
 end
 
+if ~om.is_solved()
+    error('opt_model.parse_soln: model not solved');
+end
+
 %% var
 ps = struct('var', om.var.parse_soln(om.soln, stash));
 
@@ -68,7 +72,8 @@ if ~isempty(ps_nli)
 end
 
 %%-----  DEPRECATED  -----
-%% stash the result, if requested
+%% if requested, stash the result directly in om.soln
+%% (they are already stashed in the soln property of each set type)
 if stash
     om.soln = nested_struct_copy(om.soln, ps, struct('copy_mode', '='));
 end