+++ mp-opt-model/lib/qps_osqp.m
@@ -78,7 +78,6 @@ function [x, f, eflag, output, lambda] = qps_osqp(H, c, A, l, u, xmin, xmax, x0,
 %       [x, f, exitflag, output] = qps_osqp(...)
 %       [x, f, exitflag, output, lambda] = qps_osqp(...)
 %
-%
 %   Example: (problem from from https://v8doc.sas.com/sashtml/iml/chap8/sect12.htm)
 %       H = [   1003.1  4.3     6.3     5.9;
 %               4.3     2.2     2.1     3.9;